+++ Cyan-Stars/Assets/Scripts/Gameplay/ChartEditor/Manager.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aacac6c9557c7434ab215de1e4d65b10
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
