+++ Cyan-Stars/Assets/Scripts/Gameplay/ChartEditor/ViewModel/ChartPackDataCoverViewModel.cs
@@ -8,6 +8,7 @@ using CyanStars.Framework.File;
 using CyanStars.Gameplay.ChartEditor.Command;
 using CyanStars.Gameplay.ChartEditor.Model;
 using CyanStars.Utils;
+using CyanStars.Gameplay.ChartEditor.Manager;
 using R3;
 using UnityEngine;
