+++ Cyan-Stars/Assets/Scripts/Gameplay/ChartEditor/ViewModel/ChartPackDataViewModel.cs
@@ -9,6 +9,7 @@ using CyanStars.Gameplay.ChartEditor.Command;
 using CyanStars.Gameplay.ChartEditor.Model;
 using CyanStars.Gameplay.ChartEditor.View;
 using CyanStars.Utils;
+using CyanStars.Gameplay.ChartEditor.Manager;
 using R3;
 using UnityEngine;
