+++ Cyan-Stars/Assets/Scripts/Gameplay/ChartEditor/ViewModel/MenuButtonsViewModel.cs
@@ -4,6 +4,7 @@ using CyanStars.Framework;
 using CyanStars.Gameplay.Base;
 using CyanStars.Gameplay.ChartEditor.Command;
 using CyanStars.Gameplay.ChartEditor.Model;
+using CyanStars.Gameplay.ChartEditor.Manager;
 using R3;
 
 namespace CyanStars.Gameplay.ChartEditor.ViewModel
